Run Viewer via Spring Boot
Modified the Dockerfile to run the viewer via Spring Boot alongside
serving the local json files.
Spring Boot should now be configured to serve the production build
of the OSCAL viewer as its static files, so the Dockerfile only
needs the .jar file that contains the Spring Boot app.
This reduces the number of dependencies/installations when building
the image and allows for everything to be run in one process, which
is more convenient for logging.

The Dockerfile also now builds off the official openjdk image, which
further simplifies the dependencies.