Override Spring Static File Location to Point to Viewer Build
Modify the Dockerfile to copy in the production build of the
OSCAL Viewer and specify an environment variable to override
the file path of the static files in Spring to point to the build.
This allows for the separation of the REST Service .jar and the
production build of the viewer within the container.