Push to GitHub Container Registry via GitHub Actions
Modified the GitHub actions to push to the GitHub Container Registry
instead of ECR per the team's descision to switch registries.

The new workflow first uses the docker/setup-buildx-action to set up
the Docker builder,
then uses the docker/build-push-action to build the image before testing.
If tests pass, the docker/login-action is used to log into the
container registry via the GITHUB_TOKEN provided by GitHub actions.
Finally the docker/build-push-action is used again to push the
cached image to the registry.