Pull From GitHub Packages in the Pipeline
Updated the pipeline GitHub actions config to pull from GitHub packages
instead of checking out several repos.
Adds a step calling mvn dependency:copy to pull from the Maven registry,
and a step calling npm install to pull from the npm registry.
Because the .jar file is now pulled directly from the Maven package,
the "build REST service" step can be removed.
Also changed the workflow to trigger on push to this branch so that
the Actions can be tested before merging.