Use PAT for Maven Authentication
Use a PAT instead of the "built-in" GITHUB_TOKEN for authentication
to the remote repository within GitHub Actions.
This change is necessary because GITHUB_TOKEN is scoped only to this
repo, and we need to pull from our other repos.
Made changes to the workflow config and to settings.xml to configure
the authentication.