Use Direct Download for Pulling npm Package
Change the method of pulling the production build from using `npm install`
to using a `curl` command to download a tarball directly.
This is a workaround for issues with the oscal-viewer package missing some
dependencies; this is being worked on separately.
In any case, this direct download approach may be more suitable for
the pipeline anyway as it avoids pulling in unecessary dependencies from npm.