Update Dockerfile to Pull Required Resources
Overhauled the Dockerfile to now automatically pull in the
OSCAL Viewer and OSCAL REST Service.
In its RUN step, it makes requests to the GitHub API and GitHub Packages
to get the latest versions and download them directly.

The Dockerfile takes in an argument, PACKAGE_VERSION, which is the
GitHub PAT used to authenticate all the requests.
So the image can be built with a command like:
docker build --build-arg PACKAGE_TOKEN={token} .

For now this change is not reflected in the workflow, but if this
approach works well the pipeline will be updated to simply run
`Docker build` and let the build command handle pulling everything.