Create Bash Script for Pulling from GitHub Packages
Moved the functionality of pulling the OSCAL Viewer and OSCAL REST Service
to a bash script, instead of within the Dockerfile. This allows developers
to manually add any version of the Viewer and REST Service to the Docker image
if they choose.

To accomodate this, added build arguments to the Dockerfile for configurable
paths to the Viewer and REST Service.

Modified the GitHub Actions workflow to call this script when it needs to pull
in the required resources.