Fix Docker Build and Push Action Config
Made changes to the Dockerfile and the workflow to properly set up
the "Docker build and push" action.
Added a "set up QEMU" step as specified in the official docs for the action.
Added a `context` input so the COPY step of the Dockerfile has access
to the necessary files.