Add Instructions to README for Pulling Container from Registry
Added a section to the all-in-one README for pulling the Docker container
from the GitHub container registry.
As part of this change, re-ordered the README to be more clear for end users
to easily set up and run the image.
Made small semantic changes to reflect that building the image manually
is now optional as pulling from the registry is likely the more convenient
option.