Refactor packages_pull
This reworks the script pretty significantly but the logic of the script
is mostly retained. The script now accepts the PAT via environment
variable in addition to as a positional argument on the command; the
usage text is updated to reflect this. The script checks for errors
after more commands and emits less verbose output, making it easier to
see where things went wrong. Additionally, the script now ensures that
the necessary commands (jq, unzip, etc) are installed.