Add Cypress Test Artifacts Step
Add a step to the workflow to upload the screenshots and videos output
by Cypress as artifacts.
This is being done as part of this branch in order to test that the
Cypress tests are properly updated to match the changes made to other
parts of the repo that caused the workflow to fail.