Use a relative URL for the REST API base
Users may attempt to access the viewer from across a network rather than
directly on their machine. Hardcoding the `OSCAL_REST_BASE_URL` to
`localhost` means that if a user accesses the application over a network
(imagine, at the IP address `192.168.1.20`), their API requests will go
to `http://localhost:8080/oscal/v1/catalogs` rather than to
`http://192.168.1.20/oscal/v1/catalogs`.

The viewer application uses the `fetch` API within browsers to make the
requests to the API. `fetch` supports querying a relative URL (and in
fact does so automatically).

The REST API service and the viewer application will be served from the
same port on the same IP. Therefore, we can use a relative path to
support queries when the container is accessed over a network.