Remove unneeded packages
There are a lot of packages in the regular dependencies that I don't
think we actually need. There isn't a real "package" here so everything
is sort of a devDependency. You could also maybe make the argument that
everything is actually a regular dependency and I'm happy to move these
there. But in any case, we have no need to depend directly on `ajv` or
any of the other hunk of packages that _were_ in here.