Support arbitrary `oscal-rest-service` branches
This makes `oscal-rest-service` work almost identically to the Viewer
application as far as the build process goes. A separate stage builds
the REST service which is then copied into the final stage. This limits
the size and amount of tooling in the final stage.

Some general refactoring was done as well. The versions of images were
parameterized. Since don't care about the Node version really, we just
pull the latest LTS so we don't have to worry about specific versions.
For the Java container itself, we now pull `-slim-bullseye` instead of
`-jre-slim` as the `-jre`-related containers are not present for newer
versions of Java and the base image after Java 11 changed to Oracle
Linux. We _do not_ change the default version of Java in use as part of
this PR, we just parameterize it.

The names of files were also changed around. Having a directory just
named `build` becomes a little confusing when there are now two built
artifacts, so at least in the final image, they're given more specific
names.

The documentation on how to build the container is significantly
reworded.