Always build viewer and service from source
We have to always build the viewer because the one provided as an
artifact on `oscal-react-library` is not built with REST Mode enabled
(as this is a static build-time value). Because of that, we have to
build it from source. For consistency, the REST service is always built
from source as well.