Build Docker container and run tests daily
We run into issues pretty regularly where changes happen in another repo
that don't necessarily get caught by the end-to-end tests for awhile
because we don't open/merge PRs on this repo every day. Ideally, we'd do
something event-based but that takes some effort to do cross-repo and
this is something small that gives us incremental improvement today.