Navigate Page Turning off REST Mode
Add the second test suite for drawer navigation in Non REST Mode

Tests if the page will successfully load page for each of the
components of drawer when turning off rest mode.

Each Test will navigate to the page with rest mode on, turn REST off
and check if the page will load its corresponding components.