Editor navigation by URLS e2e test
One of the many ways to navigate to the Editor views of the
drawer components is to navogate by URLs according to the
uuid (unique tag at the end of the URL).

These sets of tests test that a user can navigate to the various
editors by changing the URL to one that matchs an existing editor.
These tests navigate from one editor page to another editor page.

This also a subtask for issue #305