Switch to using distroless image
In order to switch to using a distroless image, the base image obviously
has to change, but a couple other items in the Dockerfile have to be
modified as well. The USER is now set to a pre-existing non-root user
and group, both named "nonroot". The ENTRYPOINT command has also been
replaced with CMD to avoid the container attempting to use /bin/sh that
doesn't exist within the image.