fix(workflows): auto-update workflow fails on cache setup
In a recent PR we deleted the `package-lock.json` (which was good
because it was not correct to have it there); however, that was what was
allowing `cache: npm` in the configuration for `setup-node`. That
shouldn't have been the way things were working because the cache key
would be sorta borked and changes wouldn't really be detected correctly.
We now use the `cache-dependency-path` configuration value that allows
specifying the path the lock file.