chore(workflows): dont write container logs to stdout (#219)
When running the CI workflow, the logs are written to stdout and also to
a file (using `tee(1)`). While this is sometimes helpful (because the
logs are visible in the Actions tab), it's usually more detrimental
(because there are too many logs). In fact, we currently have so many
logs that GitHub Actions wastes 3.5min just writing out logs, and that
slows down the process for getting feedback from CI.