chore(workflows): avoid building the container twice
It seems like the workflow is not using the cache from the first build
and is instead building and pushing again on the `push` step. Hopefully
reusing the tabs and labels between the jobs helps with this? Let's find
out!

Note that a full test requires a `push` event on the `main` branch in
order to work because of how the workflow works. So we need to keep an
eye on the merge to `main` to ensure this works properly (but we also
should not see regressions in the PR).