chore(workflow): refactor container image workflow
This refactors the whole workflow a bit. It splits the deployment steps
into a separate workflow with fewer triggers. This removes the
`if: ${{github.event_name == 'pull_request' }}` checks that were all
over the place and made it harder to reason about when things ran. It
also ensures that no matter what, these steps can only be run from a
more trusted branch (`main` and tags) while the more flexible testing
workflow no longer has the deploy step anywhere in it and means we don't
have to be quite as strict about the checks on the tags.

This should also incorporate the goals from #224 and avoid building the
container image twice when we run on the `main` branch. This likely
won't make things much faster for PRs but hopefully it's less confusing.