fix: clicking twice causes test failure
For some reason the first click when selecting catalogs in the
drawer selector fails. This leads to A LOT of tests failing as we use
this function in pretty much all our tests.

This temporary fix will let us test other functionality rather than just
always failing. We really should completely rework how we naviagte to each
document type in the future as this code is extemely fragile.