fix: tests failed to open catalog list (#261)
### Navigation Failing

For some reason the first click when selecting catalogs in the drawer
selector fails. This leads to A LOT of tests failing as we use this
function in pretty much all our tests.

This temporary fix will let us test other functionality rather than just
always failing. We really should **completely rework** how we navigate
to each document type in the future as this code is extremely fragile.

### Loading Invalid Components Failing

When we intentionally put in invalid data, a `TypeError` is thrown. We
are still able to reload new data after that failure, so we catch the
exception. This test may be want to reconsidered when
https://github.com/EasyDynamics/oscal-react-library/issues/872 or
subsequent related issues are completed.